Added ruby test for creating a tree of nodes trough json. Used double instead of long for storing numerical data.

+++ slingtests/osgikernel/bundles/batch/src/main/java/org/sakaiproject/kernel/batch/CreateTreeOperation.java
@@ -125,7 +125,7 @@ public class CreateTreeOperation extends AbstractSlingPostOperation {
             node.setProperty(key, (String) obj);
           }
           if (obj instanceof Number) {
-            node.setProperty(key, json.getLong(key));
+            node.setProperty(key, json.getDouble(key));
           }
           if (obj instanceof Boolean) {
             node.setProperty(key, json.getBoolean(key));
